requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.23
pydantic-settings>=2.0.0
langchain>=0.1.16
langchain-community>=0.0.26
openai>=1.0.0
psycopg2-binary>=2.9.1
pytest>=6.2.5
httpx>=0.18.2
python-multipart>=0.0.5
pandas>=1.3.3
numpy>=1.21.2
pymongo>=4.0.0 
tiktoken>=0.3.1
requests>=2.26.0
langchain_openai
python-telegram-bot